Remove temp SysteV2 file
