+++ classes/System.php
@@ -1,101 +1,33 @@
 <?php
+require_once __DIR__ . '/../classes/_sysRequires.php';
+class System {
+    //TODO: Move these
+    const KUNAI_PER_DOLLAR = 2; // Currnecy
+    const CURRENCY_TYPE_MONEY = 'money'; // Currency
+    const CURRENCY_TYPE_PREMIIUM_CREDITS = 'premium_credits'; // Currency
+    const BLOODLINE_ROLL_CHANCE = 50; // Bloodline
+    const ARENA_COOLDOWN = 4 * 1000; // Battle??
+    const DEFAULT_LAYOUT = 'new_geisha'; // Layout
+    const EXTENDED_BOOST_MULTIPLIER = 5; // Multiplies long by this value to calc extend (TrainingManager)
+    const SC_MAX_RANK = 4; // RankManager
 
-require_once __DIR__ . '/Database.php';
-require_once __DIR__ . '/EntityId.php';
-require_once __DIR__ . '/User.php';
-require_once __DIR__ . '/MarkdownParser.php';
-require_once __DIR__ . '/API.php';
-require_once __DIR__ . '/Layout.php';
-require_once __DIR__ . '/Router.php';
-require_once __DIR__ . '/Route.php';
-require_once __DIR__ . '/../classes/event/DoubleExpEvent.php';
-require_once __DIR__ . '/../classes/event/DoubleReputationEvent.php';
-require_once __DIR__ . '/../classes/event/BonusExpWeekend.php';
-require_once __DIR__ . '/../classes/event/HolidayBonusEvent.php';
+    public static array $villages = ['Stone', 'Cloud', 'Leaf', 'Sand', 'Mist']; // Village
 
-/*	Class:		System
-	Purpose: 	Handle database connection and queries. Handle storing and printing of error messages.
-*/
-class SystemOld {
-    const ENVIRONMENT_DEV = 'dev';
-    const ENABLE_DEV_ONLY_FEATURES = true;
-    const ENVIRONMENT_PROD = 'prod';
-    const LOCAL_HOST = true;
-
-    const MAINTENANCE_BEGIN = 'Jan 1 00:00 2024';
-    const MAINTENANCE_END = 'Jan 1 00:00 2024';
-
-    const KUNAI_PER_DOLLAR = 2;
-    const LOGOUT_LIMIT = 720;
-    const BLOODLINE_ROLL_CHANCE = 50;
-    const ARENA_COOLDOWN = 4 * 1000;
-
-    const CURRENCY_TYPE_MONEY = 'money';
-    const CURRENCY_TYPE_PREMIUM_CREDITS = 'premium_credits';
-
-    const SC_ADMIN_EMAIL = "admin@shinobichronicles.com";
-    const SC_NO_REPLY_EMAIL = "no-reply@shinobichronicles.com";
-    const UNSERVICEABLE_EMAIL_DOMAINS = ['hotmail.com', 'live.com', 'msn.com', 'outlook.com'];
-
+    // Event
     const HOLIDAYS = [
-		'Jan 1' => 'New Years Boost',
-	    'Jan 16' => 'Martin Luther King Day Boost',
-	    'Feb 14' => 'Valentines Day Boost',
-	    'May 27' => 'Memorial Day Boost',
-	    'Jun 19' => 'Juneteenth Boost',
-	    'Jul 4' => 'Independence Day Boost',
-	    'Sep 2' => 'Labor Day Boost',
-	    'Nov 11' => 'Veterans Day Boost',
-	    'Nov 28' => 'Thanksgiving Day Boost',
-	    'Dec 25' => 'Christmas Day Boost',
-	];
+        'Jan 1' => 'New Years Boost',
+        'Jan 15' => 'Martin Luther King Day Boost',
+        'Feb 14' => 'Valentines Day Boost',
+        'May 27' => 'Memorial Day Boost',
+        'Jun 19' => 'Juneteenth Boost',
+        'Jul 4' => 'Independence Day Boost',
+        'Sep 2' => 'Labor Day Boost',
+        'Nov 11' => 'Veterans Day Boost',
+        'Nov 28' => 'Thanksgiving Day Boost',
+        'Dec 25' => 'Christmas Day Boost',
+    ];
 
-    // Temporary event data storage
-    public ?Event $event;
-
-    public static array $villages = ['Stone', 'Cloud', 'Leaf', 'Sand', 'Mist'];
-
-    // Variable for error message
-    public string $message = "";
-    public bool $message_displayed = false;
-
-    public array $debug_messages = [];
-
-    // Sub-components
-    public Database $db;
-    public Router $router;
-    public ?Layout $layout;
-    public bool $enable_mobile_layout = false;
-
-    public string $environment;
-    public bool $enable_dev_only_features;
-
-    public bool $SC_OPEN;
-    public ?DateTimeImmutable $UPDATE_MAINTENANCE;
-    public bool $register_open;
-    public bool $USE_NEW_BATTLES = false;
-
-    public string $link;
-
-    // Request lifecycle
-    public bool $is_api_request = false;
-    public bool $is_legacy_ajax_request = false;
-
-    //New server time
-    const SERVER_TIME_ZONE = "America/New_York";
-    const REPUTATION_RESET_DAY = "Friday";
-    const REPUTATION_RESET_HOUR = 20;
-    const REPUTATION_RESET_MINUTE = 0;
-    public DateTimeImmutable $SERVER_TIME;
-    public DateTimeImmutable $REPUTATION_RESET;
-    // Old server time
-    public $timezoneOffset;
-
-    // Training boost switches
-    public int $TRAIN_BOOST = 0; // Extra points per training, 0 for none
-    public int $LONG_TRAIN_BOOST = 0; // Extra points per long training, 0 for none
-    const EXTENDED_BOOST_MULTIPLIER = 5; // Multiples long by this value to calc extended boost
-
+    // StaffManager, handle differently with CSS overhaul
     public array $SC_STAFF_COLORS = array(
         User::STAFF_MODERATOR => array(
             'staffBanner' => "moderator",
@@ -124,24 +56,7 @@ class SystemOld {
         )
     );
 
-    // Default layout
-    const DEFAULT_LAYOUT = 'new_geisha';
-    const VERSION_NUMBER = '0.11';
-    const VERSION_NAME = '0.11 Warring Shadows';
-
-    // Misc stuff
-    const SC_MAX_RANK = 4;
-
-    const MAX_LINK_DISPLAY_LENGTH = 60;
-
-    public bool $war_enabled = false;
-
-    // Currently this only enables errors on production
-    public static array $developers = [
-        1, // Lsmjudoka
-        254, // Hitori
-        1603, // Arth
-    ];
+    // TEMPORARY - Need to find proper parser
     public static array $explicit_words = [
         'fuck',
         'fuk',
@@ -207,93 +122,175 @@ class SystemOld {
         'nigger',
     ];
 
-    public array $debug = [
-        'battle' => false,
-        'battle_effects' => false,
-        'jutsu_collision' => false,
-        'damage' => false,
-        'bloodline' => false,
-        'stat_cut' => false,
+    // END TODO
+
+    // Environment settings
+    const ENVIRONMENT_DEV = 'dev';
+    const ENVIRONMENT_PROD = 'prod';
+    const DEV_ONLY_FEATURES_DEFAULT = false;
+    const LOCAL_HOST = true;
+
+    const VERSION_NUMBER = '0.11';
+    const VERSION_NAME = '0.11 Warring Shadows';
+
+    const LOGOUT_LIMIT = 720;
+    const MAX_LINK_DISPLAY_LENGTH = 60;
+
+    const SC_ADMIN_EMAIL = "admin@shinobichronicles.com";
+    const SC_NO_REPLY_EMAIL = "no-reply@shinobichronicles.com";
+    const UNSERVICEABLE_EMAIL_DOMAINS = ['hotmail.com', 'live.com', 'msn.com', 'outlook.com'];
+
+    // Currently this only enables errors on production
+    public static array $developers = [
+        1, // Lsmjudoka
+        254, // Hitori
+        1603, // Arth
     ];
 
-    const GLOBAL_TEST_NOTIFICATION_SPAWN = false;
-    public array $testNotifications = [
-        'caravan' => self::GLOBAL_TEST_NOTIFICATION_SPAWN,
-        'raid' => self::GLOBAL_TEST_NOTIFICATION_SPAWN,
-        'event' => self::GLOBAL_TEST_NOTIFICATION_SPAWN,
-        'diplomacy' => self::GLOBAL_TEST_NOTIFICATION_SPAWN,
-    ];
-    public bool $REQUIRE_USER_VERIFICATION;
+    // Time-based settings
+    const SERVER_TIME_ZONE = 'America/New_York';
+    const REPUTATION_RESET_DAY = 'Friday';
+    const REPUTATION_RESET_HOUR = 20;
+    const REPUTATION_RESET_MINUTE = 0;
 
-    public function __construct() {
-        require __DIR__ . "/../secure/vars.php";
-        /** @var $host */
-        /** @var $username */
-        /** @var $password */
-        /** @var $database */
-        $this->db = new Database($host, $username, $password, $database);
+    public function __construct(
+        public Database $db,
+        public Router $router,
+        public bool $SC_OPEN,
+        public readonly bool $USE_NEW_BATTLES,
+        public readonly bool $war_enabled,
+        public readonly bool $REQUIRE_USER_VERIFICATION,
+        public DateTimeImmutable $SERVER_TIME,
+        public ?DateTimeImmutable $UPDATE_MAINTENANCE = null,
+        public ?DateTimeImmutable $REPUTATION_RESET = null,
+        public ?Layout $layout = null,
+        public ?int $timezoneOffset = null,
+        public bool $enable_mobile_layout = false,
+        public string $environment = self::ENVIRONMENT_DEV,
+        public readonly bool $enable_dev_only_features = self::DEV_ONLY_FEATURES_DEFAULT,
+        public readonly bool $local_host = self::LOCAL_HOST,
+        public readonly bool $register_open = false,
+        public ?Event $event = null,
+        public string $message = '',
+        public bool $message_displayed = false,
+        public array $debug_messages = [],
+        public bool $is_api_request = false,
+        public bool $is_legacy_ajax_request = false,
+        public int $TRAIN_BOOST = 0, // Extra points per training - REMOVE? Handle with TrainingManager & Events
+        public int $LONG_TRAIN_BOOST = 0, // Extra points per long training - Remave as above?
+        public readonly array $debug = [
+            'battle' => false,
+            'battle_effects' => false,
+            'jutsu_collision' => false,
+            'damage' => false,
+            'bloodline' => false,
+            'stat_cut' => false,
+        ],
+        public readonly array $testNotifications = [
+            'caravan' => false,
+            'raid' => false,
+            'event' => false,
+            'diplomacy' => false,
+        ],
+        public array $homeVars = [],
+    ){}
 
-        $this->environment = $ENVIRONMENT ?? self::ENVIRONMENT_DEV;
-        $this->enable_dev_only_features = $ENABLE_DEV_ONLY_FEATURES ?? self::ENABLE_DEV_ONLY_FEATURES; // Will only ever effect dev envs
-        $this->register_open = $register_open ?? false;
-        $this->SC_OPEN = $SC_OPEN ?? false;
-        $this->SERVER_TIME = new DateTimeImmutable(datetime: "now", timezone: new DateTimeZone(self::SERVER_TIME_ZONE));
-        $this->loadRepReset();
-        $this->UPDATE_MAINTENANCE = $this->checkForMaintenance();
-        $this->USE_NEW_BATTLES = $USE_NEW_BATTLES ?? false;
-
-        $this->router = new Router($web_url ?? 'http://localhost/');
-
-        // Old Server Time
-        $this->timezoneOffset = date('Z');
-
-        $this->checkForActiveEvent();
-
-        $this->war_enabled = true;
-        // Temp verification fix to require users to validate email once we can send those again
-        $this->REQUIRE_USER_VERIFICATION = false;
-    }
+    /******************************************************
+     *        Debuging, System Messages and Strings       *
+     ******************************************************/
 
     /**
-     * @param $message
-     * @param bool $force_message
+     * Sets message for display in UI
+     * @param $message - String value to be displayed on UI
+     * @param bool $force_message - Allows overriding previous message, if not displayed
      * @return void
      */
-    public function message($message, bool $force_message = false): void {
-        if(strlen($this->message) == 0 || $force_message) {
+    public function message(string $message, bool $force_message = false): void {
+        if(!strlen($this->message) || $force_message) {
             $this->message = $message;
         }
     }
 
-    public function debugMessage($message): void {
-        $this->debug_messages[] = $message;
-    }
-
-    /* function printMessage()
-        Displays message, if one is stored.
-        -Parameters-
-        None
-    */
-    public function printMessage($force_display = false): bool {
+    /**
+     * Displays system message, if one is set
+     * TODO: Can this be a void method? Make a more modern display method
+     * @param bool $force_display - Displays message regardless of length or previous display
+     * @return bool
+     */
+    public function printMessage(bool $force_display = false): bool {
         if(strlen($this->message) && (!$this->message_displayed || $force_display)) {
             echo "<p class='systemMessage'>{$this->message}</p>";
             $this->message = '';
             $this->message_displayed = true;
-            return true;
         }
         return true;
     }
 
-    public static function currentTimeMs(): int {
-        return floor(microtime(true) * 1000);
+    /**
+     * Fetches global message string and time and returns as an array
+     * @return ?array
+     */
+    public function fetchGlobalMessage(): ?array {
+        $result = $this->db->query("SELECT `global_message`, `time` FROM `system_storage` LIMIT 1");
+        if($result->num_rows < 1) {
+            return null;
+        }
+
+        $results = $this->db->fetch($result);
+
+        return [
+            'message' => str_replace("\r\n", "<br />", $results['global_message']),
+            'time' => date("l, M j, Y - g:i A", $results['time'])
+        ];
     }
 
-    /* function explicitLanguageCheck(word)
-        Checks our list of banned words, returns true or false if censored words are detected.
-        -Parameters-
-        @string
-    */
-    public function explicitLanguageCheck($string): bool {
+    /**
+     * Stores debugging messages
+     * @return void
+     */
+    public function debugMessage($message): void {
+        $this->debug_messages[] = $message;
+    }
+
+    /**
+     *
+     * @return Closure
+     */
+    public static function getEchoDebugClosure(): Closure {
+        return function ($category, $label, $contents) {
+            if(php_sapi_name() == "cli") {
+                echo "\r\nDEBUG ($label)\r\n" . $contents . "\r\n";
+            }
+            else {
+                echo "<br />DEBUG ($label)<br />" . $contents . "<br />";
+            }
+        };
+    }
+
+    /**
+     *
+     * @return string
+     */
+    public static function simpleStackTrace(): string {
+        $stack_trace_arr = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS, 4);
+
+        // Display the first line as file_name (line # would be the line that called this method, not super useful)
+        $display = $stack_trace_arr[0]['file'];
+
+        // Display the rest of the lines as file:line_num
+        foreach(array_slice($stack_trace_arr, 1) as $stack_trace_row) {
+            $display .= " > " . $stack_trace_row['file'] . ":" . $stack_trace_row['line'];
+        }
+
+        return $display;
+    }
+
+    /**
+     * Checks string for banned words and returns bool based on check
+     * @param string $string
+     * @return bool
+     */
+    public function explicitLanguageCheck(string $string): bool {
         foreach(self::$explicit_words as $word) {
             if(str_contains(strtolower($string), $word)) {
                 return true;
@@ -303,13 +300,14 @@ class SystemOld {
     }
 
     /**
-     * @param      $text
+     * Returns censored explicit words from string, based on user settings
+     * Banned words are censored regardless of user settings
+     * @param string $text
      * @param bool $allow_non_banned_words
      * @return string
      */
-    public function explicitLanguageReplace($text, bool $allow_non_banned_words = false): string {
+    public function explicitLanguageReplace(string $text, bool $allow_non_banned_words): string {
         $word_replace = [];
-
         if(!$allow_non_banned_words) {
             foreach(self::$explicit_words as $word) {
                 $word_replace[] = str_repeat('*', strlen($word));
@@ -318,21 +316,16 @@ class SystemOld {
             $text = str_ireplace(self::$explicit_words, $word_replace, $text);
         }
 
-
-        $word_replace = [];
-        foreach(self::$banned_words as $word) {
-            $word_replace[] = str_repeat('*', strlen($word));
-        }
-
-        return str_ireplace(self::$banned_words, $word_replace, $text);
+        // Censor banned words, regardless of user settings
+        return $this->bannedLanguageReplace($text);
     }
 
     /**
-     * @param      $text
+     * Retruns censored banned words from string, regardless of user settings
+     * @param string $text
      * @return string
      */
-    public function bannedLanguageReplace($text): string {
-        //String censorship
+    public function bannedLanguageReplace(string $text): string {
         $word_replace = [];
         foreach(self::$banned_words as $key => $word) {
             $word_replace[] = str_repeat('*', strlen($word));
@@ -341,46 +334,65 @@ class SystemOld {
         return str_ireplace(self::$banned_words, $word_replace, $text);
     }
 
+    /**
+     * Fetches and returns array of memes based on meme file
+     * @return array
+     */
     public function getMemes(): array {
-        $memes_dir = __DIR__ . '/../images/memes';
-        $meme_files = scandir($memes_dir);
-        $file_type_filter = '/(?:.png|.jpg|.gif)$/i';
+        $meme_dir = __DIR__ . '/../images/memes';
+        $meme_files = scandir($meme_dir);
 
-        if ($meme_files === False)
-        {
-            return [];
+        // Memes not found
+        if(!$meme_files) {
+            return array();
         }
-        $cleaned_memes = array_filter($meme_files, function ($meme) use ($file_type_filter) {
-            return preg_match($file_type_filter, $meme) == 1;
-        });
+
+        // Regex to filter out directories and non-image files
+        $file_type_filter = '/(?:.png|.jpg|.gif)$/i';
         $search_symbols = ['-', '_'];
         $meme_array = [
             'codes' => [],
             'images' => [],
             'urls' => [],
-            'texts' => []
+            'texts' => [],
         ];
-        foreach ($cleaned_memes as $meme)
-        {
-            $meme_code = strtolower(':' . preg_replace($file_type_filter,'', str_replace($search_symbols, '', $meme)) . ':');
 
-            if (in_array($meme_code, $meme_array['codes'])) continue;
+        // Set memes for storage in array
+        $cleaned_memes = array_filter($meme_files, function($meme) use ($file_type_filter) {
+            return preg_match($file_type_filter, $meme) == 1;
+        });
 
-            $url = "./images/memes/${meme}";
+        foreach($cleaned_memes as $meme) {
+            $meme_code = strtolower(':' . preg_replace($file_type_filter, '', str_replace($search_symbols, '', $meme)) . ':');
+            $url = "./images/memes/$meme";
+
+            // Skip duplicate meme
+            if(in_array($meme_code, $meme_array['codes'])) {
+                continue;
+            }
 
             $meme_array['codes'][] = $meme_code;
-            $meme_array['images'][] = "<img src='{$url}' title='${meme_code}' alt='${meme_code}' style='max-width: 75px;max-height: 75px'/>";
+            $meme_array['images'][] = "<img src='$url' title='$meme_code' alt='$meme_code' style='max_width: 75px; max-height: 75px;' />";
             $meme_array['urls'][] = $url;
             $meme_array['texts'][] = $meme_code;
         }
+
         return $meme_array;
     }
 
-    public function html_parse($text, $img = false, $faces = false, $youtube = false): array|string {
+    /**
+     * Parses string for html/bbcode, images, memes, and youtube links
+     * @var string $text
+     * @var bool $img
+     * @var bool $faces
+     * @var bool $youtube
+     * @return array|string
+     */
+    public function html_parse(string $text, bool $img = false, bool $faces = false, $youtube = false): array|string {
         $search_array = array(
             "[b]","[/b]","[u]","[/u]","[i]","[/i]",
             "&lt;3","[strike]","[/strike]","[super]","[/super]","[sub]","[/sub]", "[center]", "[/center]", "[right]", "[/right]",
-            );
+        );
         $replace_array = array("<b>","</b>","<u>","</u>","<i>","</i>","&hearts;",
             "<del>","</del>","<sup>","</sup>","<sub>","</sub>", "<p style='text-align:center;'>", "</p>",
             "<p style='text-align:right;'>", "</p>",
@@ -471,6 +483,14 @@ class SystemOld {
         return $text;
     }
 
+    /**
+     * Returns formatted string from MarkdownParser
+     * @var string $text
+     * @var bool $allow_images
+     * @var bool $strip_breaks
+     * @var bool $faces
+     * @return string
+     */
     public function parseMarkdown($text, $allow_images = false, $strip_breaks = true, $faces = false): string {
         if($strip_breaks) {
             $text = str_replace("\n", "", $text);
@@ -484,16 +504,124 @@ class SystemOld {
             ->text($text);
     }
 
-    public function imageCheck($image, $size): string {
-        $avatar_limit = $size;
-
-        $width = $avatar_limit;
-        $height = $avatar_limit;
-
-        return "<img src='$image' style='max-width:{$width}px;max-height:{$height}px;' />";
+    /**
+     * Returns img markup, intended for user avatars
+     * @var string $image_src
+     * @var int $size
+     * @return string
+     */
+    public function imageCheck(string $image_src, int $size): string {
+        return "<img src='$image_src' style='max-width:{$size}px;max-height:{$size}px;' />";
     }
 
-    public function timeAgo($timestamp): string {
+    /**
+     * Formats readable variables into db-friendly string for storage
+     * @var string $string
+     * @return string
+     */
+    public static function slug(string $string): string {
+        return strtolower(str_replace(' ', '_', $string));
+    }
+    /**
+     * Formats named values from db for display
+     * @var string $slug
+     * @return string
+     */
+    public static function unSlug(string $slug): string {
+        return ucwords(str_replace('_', ' ', $slug));
+    }
+
+    /******************************************************
+     *                    TIME FUNCTIONS                  *
+     ******************************************************/
+
+    /**
+     * Uses server time to determine reputation reset time.
+     * This must be store as DateTimeImmutable.
+     * @return void
+     */
+    public function loadRepReset(): void {
+        // Reset is today
+        if(strtolower($this->SERVER_TIME->format('l')) == strtolower(self::REPUTATION_RESET_DAY)) {
+            // Set time to today at proper hour and minute
+            $this->REPUTATION_RESET = $this->SERVER_TIME->setTime(hour: self::REPUTATION_RESET_HOUR, minute: self::REPUTATION_RESET_MINUTE);
+            // Reset has passed for today, move to next week and set to proper hour & minute
+            if($this->REPUTATION_RESET->getTimestamp() <= $this->SERVER_TIME->getTimestamp()) {
+                $this->REPUTATION_RESET = $this->SERVER_TIME->modify('next ' . self::REPUTATION_RESET_DAY);
+                $this->REPUTATION_RESET = $this->REPUTATION_RESET->setTime(hour: self::REPUTATION_RESET_HOUR, minute: self::REPUTATION_RESET_MINUTE);
+            }
+        }
+        // Reset is later in the week
+        else {
+            $this->REPUTATION_RESET = $this->SERVER_TIME->modify('next ' . self::REPUTATION_RESET_DAY);
+            $this->REPUTATION_RESET = $this->REPUTATION_RESET->setTime(hour: self::REPUTATION_RESET_HOUR, minute: self::REPUTATION_RESET_MINUTE);
+        }
+    }
+
+    /**
+     * Returns current time in rounded miliseconds
+     * @return int
+     */
+    public static function currentTimeMs(): int {
+        return floor(microtime(true) * 1000);
+    }
+
+    /**
+     * Returns DateTime based on provided microtime
+     * @var float $microtime
+     * @return DateTime|bool
+     */
+    public static function dateTimeFromMicrotime(float $microtime): DateTime|bool {
+        return DateTime::createFromFormat(
+            'U.u',
+            number_format($microtime, 2, '.', '')
+        );
+    }
+
+    public static function currentYear(): int {
+        return (int) date('Y', time());
+    }
+
+    /**
+     * @return int month of the year: 1-12
+     */
+    public static function currentMonth(): int {
+        return (int) date('n', time());
+    }
+
+    /**
+     * @return int day of the month: 1-31
+     */
+    public static function currentDay(): int {
+        return (int) date('j', time());
+    }
+
+    /**
+     * @return int hour of the day in 24-hour format: 0-23
+     */
+    public static function currentHour(): int {
+        return (int) date('G', time());
+    }
+
+    /**
+     * @return int minute of the hour, 0-59
+     */
+    public static function currentMinute(): int {
+        return (int)date('i');
+    }
+
+    /**
+     * @return int day of the week: 0 (Sunday) - 6 (Saturday)
+     */
+    public static function currentDayOfWeek(): int {
+        return (int) date('w');
+    }
+
+    /**
+     * Returns formatted string based on a previous unix timestamp
+     * @var int $timestamp
+     */
+    public function timeAgo(int $timestamp): string {
         $time = time() - $timestamp;
 
         $days = 0;
@@ -532,7 +660,12 @@ class SystemOld {
         return $time_string;
     }
 
-    public function time_remaining($timestamp): string {
+    /**
+     * Returns formatted string based on unix timestamp in the future
+     * @var int $timestamp
+     * @return string
+     */
+    public function time_remaining(int $timestamp): string {
         $days = floor($timestamp / 86400);
         $hours = floor($timestamp / 3600);
         $minutes = ceil($timestamp / 60);
@@ -556,274 +689,15 @@ class SystemOld {
         return $message;
     }
 
-    public function log($type, $title, $contents): void {
-        $type = $this->db->clean($type);
-        $title = $this->db->clean($title);
-
-        if (is_array($contents))
-        {
-            $contents = json_encode($contents);
-        }
-
-        $contents = $this->db->clean($contents);
-
-        $this->db->query(
-            "INSERT INTO `logs` (`log_type`, `log_title`, `log_time`, `log_contents`)
-                VALUES ('$type', '$title', " . time() . ", '$contents')"
-        );
-    }
-
     /**
-     * @throws RuntimeException
+     * Static method of returning formatted time string
+     * TODO: Consolidate timeRemaining functions to reduce duplicated code
+     * @var int $time_remaining
+     * @var string $format
+     * @var bool $include_days
+     * @var bool $include_seconds
+     * @return string
      */
-    public function currencyLog(
-        int $character_id,
-        string $currency_type,
-        int $previous_balance,
-        int $new_balance,
-        int $transaction_amount,
-        string $transaction_description
-    ): void {
-       switch($currency_type) {
-           case 'premium_credits':
-           case 'money':
-               break;
-           default:
-               throw new RuntimeException("Invalid currency type!");
-       }
-
-        $this->db->query(
-            "INSERT INTO `currency_logs` (
-                `character_id`,
-                `currency_type`,
-                `previous_balance`,
-                `new_balance`,
-                `transaction_amount`,
-                `transaction_description`,
-                `transaction_time`
-            ) VALUES (
-               '{$character_id}',
-                '{$currency_type}',
-                '{$previous_balance}',
-                '{$new_balance}',
-                '{$transaction_amount}',
-                '{$this->db->clean($transaction_description)}',
-                " . time() . "
-            )"
-        );
-    }
-
-    public function hash_password($password): string {
-        return password_hash($password, PASSWORD_DEFAULT);
-    }
-
-    public function verify_password($password, $hash): bool {
-        return password_verify($password, $hash);
-    }
-
-    public function setLayoutByName(string $layout): Layout {
-        $system = $this;
-
-        switch($layout) {
-            case 'cextralite':
-                $this->layout = require "layout/cextralite.php";
-                break;
-            case 'classic_blue':
-                $this->layout = require  "layout/classic_blue.php";
-                break;
-            case 'shadow_ribbon':
-                $this->layout = require  "layout/shadow_ribbon.php";
-                break;
-            case 'geisha':
-                $this->layout = require  "layout/geisha.php";
-                break;
-            case 'blue_scroll':
-                $this->layout = require  "layout/blue_scroll.php";
-                break;
-            case 'rainbow_road':
-                $this->layout = require "layout/rainbow_road.php";
-                break;
-            case 'new_geisha':
-                require_once __DIR__ . "/../layout/new_geisha.php";
-                // This needs to be first so the function can read it
-                $this->enable_mobile_layout = true;
-                $this->layout = getNewGeishaLayout($this, $this->enable_mobile_layout);
-                break;
-            case 'sumu':
-                require_once __DIR__ . "/../layout/sumu.php";
-                // This needs to be first so the function can read it
-                $this->enable_mobile_layout = true;
-                $this->layout = getSumuLayout($this, $this->enable_mobile_layout);
-                break;
-            default:
-                $this->layout = require "layout/" . self::DEFAULT_LAYOUT . ".php";
-                break;
-        }
-
-        return $this->layout;
-    }
-
-    /**
-     * @throws RuntimeException
-     */
-    public function fetchGlobalMessage(): ?array {
-        $result = $this->db->query("SELECT `global_message`, `time` FROM `system_storage` LIMIT 1");
-        if($result->num_rows < 1) {
-            return null;
-        }
-
-        $results = $this->db->fetch($result);
-
-        return [
-            'message' => str_replace("\r\n", "<br />", $results['global_message']),
-            'time' => date("l, M j, Y - g:i A", $results['time'])
-        ];
-    }
-
-    public function isDevEnvironment(): bool {
-        return $this->environment == System::ENVIRONMENT_DEV;
-    }
-
-    // Note: The system can currently only support one event type
-    public function checkForActiveEvent(): void {
-        // Base data
-        $current_datetime = new DateTimeImmutable();
-        $this->event = null;
-
-        // Dev Environment Event start times
-        if($this->isDevEnvironment()) {
-            $july_2023_lantern_event_start_time = new DateTimeImmutable('2023-07-15');
-            $double_exp_start_time = new DateTimeImmutable('2023-12-25');
-            $double_reputation_start_time = new DateTimeImmutable('2023-10-16');
-        }
-        // Production Event start times
-        else {
-            $july_2023_lantern_event_start_time = new DateTimeImmutable('2023-07-01');
-            $double_exp_start_time = new DateTimeImmutable('2023-12-25');
-            $double_reputation_start_time = new DateTimeImmutable('2023-10-18');
-        }
-
-        /*****CORE EVENTS*****/
-        // TODO: Make core events more manageable
-        // Double exp gains
-        $double_exp_end_time = new DateTimeImmutable('2023-12-27');
-        if($current_datetime > $double_exp_start_time && $current_datetime < $double_exp_end_time) {
-            $this->event = new DoubleExpEvent($double_exp_end_time);
-        }
-        // Double reputation gains
-        $double_reputation_end_time = new DateTimeImmutable('2023-10-30');
-        if($current_datetime > $double_reputation_start_time && $current_datetime < $double_reputation_end_time) {
-            $this->event = new DoubleReputationEvent($double_reputation_end_time);
-        }
-
-        /*****LIMITED TIME EVENTS*****/
-        // July 2023 Lantern Event
-        $july_2023_lantern_event_end_time = new DateTimeImmutable('2023-07-16');
-        if($current_datetime > $july_2023_lantern_event_start_time && $current_datetime < $july_2023_lantern_event_end_time) {
-            $this->event = new LanternEvent($july_2023_lantern_event_end_time);
-        }
-
-        if($this->testNotifications['event'] && is_null($this->event) && $this->isDevEnvironment()) {
-            $this->event = new DoubleExpEvent($current_datetime->modify("+2 weeks"));
-        }
-
-        // If no other event active on weekend, activate bonus XP
-        if (!isset($this->event) && (System::currentDayOfWeek() == 0 || System::currentDayOfWeek() == 6)) {
-            $endTime = new DateTimeImmutable('next Monday');
-            $this->event = new BonusExpWeekend($endTime);
-        }
-    	// Holiday training boosts, if no other event has been planned
-    	if (!isset($this->event) && in_array($current_datetime->format('M j'), array_keys(self::HOLIDAYS))) {
-		    $endTime = new DateTimeImmutable("tomorrow");
-		    $this->event = new HolidayBonusEvent($endTime, self::HOLIDAYS[$current_datetime->format('M j')]);
-	    }
-    }
-
-    public function checkForMaintenance(): ?DateTimeImmutable {
-        // Check for db-based maintenance
-        $db_maint = $this->db->query("SELECT `maintenance_begin_time`, `maintenance_end_time` FROM `system_storage`");
-        $db_maint = $this->db->fetch($db_maint);
-
-        if($db_maint['maintenance_begin_time'] > $this->SERVER_TIME->getTimestamp() || $db_maint['maintenance_end_time'] > $this->SERVER_TIME->getTimestamp()) {
-            $maintenanceBegin = $this->SERVER_TIME->setTimestamp($db_maint['maintenance_begin_time']);
-            $maintenanceEnd = $this->SERVER_TIME->setTimestamp($db_maint['maintenance_end_time']);
-        }
-        // Check for manual code maintenance
-        else {
-            $maintenanceBegin = new DateTimeImmutable(self::MAINTENANCE_BEGIN, new DateTimeZone(self::SERVER_TIME_ZONE));
-            $maintenanceEnd = new DateTimeImmutable(self::MAINTENANCE_END, new DateTimeZone(self::SERVER_TIME_ZONE));
-        }
-		
-		// Hard closure
-		if($db_maint['maintenance_end_time'] == -1) {
-			$this->SC_OPEN = false;
-			return null;
-		}
-
-        // Maintenance time-frame error
-        if($maintenanceBegin->getTimestamp() > $maintenanceEnd->getTimestamp()) {
-            return null;
-        }
-
-        // Display timer for maintenance window
-        if($this->SERVER_TIME->getTimestamp() < $maintenanceBegin->getTimestamp()) {
-            return $maintenanceBegin;
-        }
-        // Close SC for maintenance window - NOTE: This can be overridden in vars.php if window can't be easily determined
-        if($this->SERVER_TIME->getTimestamp() > $maintenanceBegin->getTimestamp() && $this->SERVER_TIME->getTimestamp() < $maintenanceEnd->getTimestamp()) {
-            $this->SC_OPEN = false;
-            return $maintenanceEnd;
-        }
-
-
-        return null;
-    }
-
-	// This will return the estimated time the server will come back online, rounded up to the nearst 5 minutes
-	// e.g. 7 minutes => 10 minutes // 3 minutes => 5 minutes
-	// Default return for closing server in secure/vars is 30 minutes
-    public function getMaintenenceEndTime(): string {
-        if(!$this->SC_OPEN && $this->UPDATE_MAINTENANCE) {
-            $mins = ceil(($this->UPDATE_MAINTENANCE->getTimestamp() - time()) / 60); // Round up to nearest minute
-            $mins += 5 - ($mins % 5); // Add remainder to bring to nearst 5 minutes
-            return "$mins minutes";
-        }
-
-        return "30 minutes";
-    }
-
-    public function loadRepReset(): void {
-		// Reset is today
-		if(strtolower($this->SERVER_TIME->format('l')) == strtolower(self::REPUTATION_RESET_DAY)) {
-			// Set time to today at proper hour and minute
-			$this->REPUTATION_RESET = $this->SERVER_TIME->setTime(hour: self::REPUTATION_RESET_HOUR, minute: self::REPUTATION_RESET_MINUTE);
-			// Reset has passed for today, move to next week and set to proper hour & minute
-			if($this->REPUTATION_RESET->getTimestamp() <= $this->SERVER_TIME->getTimestamp()) {
-				$this->REPUTATION_RESET = $this->SERVER_TIME->modify('next ' . self::REPUTATION_RESET_DAY);
-				$this->REPUTATION_RESET = $this->REPUTATION_RESET->setTime(hour: self::REPUTATION_RESET_HOUR, minute: self::REPUTATION_RESET_MINUTE);
-			}
-		}
-		// Reset is later in the week
-		else {
-			$this->REPUTATION_RESET = $this->SERVER_TIME->modify('next ' . self::REPUTATION_RESET_DAY);
-			$this->REPUTATION_RESET = $this->REPUTATION_RESET->setTime(hour: self::REPUTATION_RESET_HOUR, minute: self::REPUTATION_RESET_MINUTE);
-		}
-    }
-
-    /**
-     * @param string $entity_id
-     * @return EntityId
-     * @throws RuntimeException
-     */
-    public static function parseEntityId(string $entity_id): EntityId {
-        $arr = explode(':', $entity_id);
-        if(count($arr) != 2) {
-            throw new RuntimeException("Invalid entity id {$entity_id}!");
-        }
-
-        return new EntityId($arr[0], (int)$arr[1]);
-    }
-
     public static function timeRemaining($time_remaining, $format = 'short', $include_days = true, $include_seconds = true): string {
         if($include_days) {
             $days = floor($time_remaining / 86400);
@@ -897,72 +771,85 @@ class SystemOld {
         return $string;
     }
 
-    public static function dateTimeFromMicrotime(float $microtime): DateTime|bool {
-        return DateTime::createFromFormat(
-            'U.u',
-            number_format($microtime, 2, '.', '')
+    /******************************************************
+     *                         Logs                       *
+     ******************************************************/
+
+    /**
+     * Inserts various log types into System logs.
+     * Custom methods should be used for other managers.
+     * @var string $type
+     * @var string $title
+     * @var array|string $contents
+     * @return void
+     */
+    public function log(string $type, string $title, array|string $contents): void {
+        $type = $this->db->clean($type);
+        $title = $this->db->clean($title);
+
+        if (is_array($contents)) {
+            $contents = json_encode($contents);
+        }
+
+        $contents = $this->db->clean($contents);
+
+        $this->db->query(
+            "INSERT INTO `logs` (`log_type`, `log_title`, `log_time`, `log_contents`)
+                VALUES ('$type', '$title', " . time() . ", '$contents')"
         );
     }
 
-    public static function slug(string $string): string {
-        return strtolower(str_replace(' ', '_', $string));
-    }
+    /**
+     * Addes currency transactions to currency log
+     * TODO: Move to Currency
+     * @var int $character_id
+     * @var string $currency_type
+     * @var int $previous_balance
+     * @var int $new_balance
+     * @var int $transaction_amount
+     * @var string $transaction_description
+     * @throws RuntimeException
+     * @return void
+     */
+    public function currencyLog(
+        int $character_id, string $currency_type,
+        int $previous_balance, int $new_balance,
+        int $transaction_amount, string $transaction_description
+    ): void {
+        switch($currency_type) {
+            case 'premium_credits':
+            case 'money':
+                break;
+            default:
+                throw new RuntimeException("Invalid currency type!");
+        }
 
-    public static function unSlug(string $slug): string {
-        return ucwords(str_replace('_', ' ', $slug));
-    }
-
-    public static function getEchoDebugClosure(): Closure {
-        return function ($category, $label, $contents) {
-            if(php_sapi_name() == "cli") {
-                echo "\r\nDEBUG ($label)\r\n" . $contents . "\r\n";
-            }
-            else {
-                echo "<br />DEBUG ($label)<br />" . $contents . "<br />";
-            }
-        };
-    }
-
-    public static function currentYear(): int
-    {
-        return (int) date('Y', time());
+        $this->db->query(
+            "INSERT INTO `currency_logs` (
+                `character_id`,
+                `currency_type`,
+                `previous_balance`,
+                `new_balance`,
+                `transaction_amount`,
+                `transaction_description`,
+                `transaction_time`
+            ) VALUES (
+               '{$character_id}',
+                '{$currency_type}',
+                '{$previous_balance}',
+                '{$new_balance}',
+                '{$transaction_amount}',
+                '{$this->db->clean($transaction_description)}',
+                " . time() . "
+            )"
+        );
     }
 
     /**
-     * @return int month of the year: 1-12
+     * Returns available premium packages
+     * TODO: Move to Currency
+     * @return array
      */
-    public static function currentMonth(): int {
-        return (int) date('n', time());
-    }
-
-    /**
-     * @return int day of the month: 1-31
-     */
-    public static function currentDay(): int {
-        return (int) date('j', time());
-    }
-
-    /**
-     * @return int hour of the day in 24-hour format: 0-23
-     */
-    public static function currentHour(): int {
-        return (int) date('G', time());
-    }
-
-    /**
-     * @return int minute of the hour, 0-59
-     */
-    public static function currentMinute(): int {
-        return (int)date('i');
-    }
-
-    /**
-     * @return int day of the week: 0 (Sunday) - 6 (Saturday)
-     */
-    public static function currentDayOfWeek(): int {
-        return (int) date('w');
-    }
-
     public static function getKunaiPacks(): array {
         $kunai_packs = [
             [
@@ -993,31 +880,230 @@ class SystemOld {
         return $kunai_packs;
     }
 
-    public static function simpleStackTrace(): string {
-        $stack_trace_arr = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS, 4);
+    /******************************************************
+     *                    Miscelaenous                    *
+     ******************************************************/
 
-        // Display the first line as file_name (line # would be the line that called this method, not super useful)
-        $display = $stack_trace_arr[0]['file'];
-
-        // Display the rest of the lines as file:line_num
-        foreach(array_slice($stack_trace_arr, 1) as $stack_trace_row) {
-            $display .= " > " . $stack_trace_row['file'] . ":" . $stack_trace_row['line'];
-        }
-
-        return $display;
+    /**
+     * Hash password for db storage
+     * TODO: Move to User
+     * @var string $password
+     * @return string
+     */
+    public function hash_password(string $password): string {
+        return password_hash($password, PASSWORD_DEFAULT);
     }
 
+    /**
+     * Verifies password based on provided password and stored has
+     * TODO: Move to User
+     * @var string $password
+     * @var string $password_hash
+     * @return bool
+     */
+    public function verify_password(string $password, string $password_hash): bool {
+        return password_verify($password, $password_hash);
+    }
+
+    /**
+     * Returns whether or not environment is dev
+     * @return bool
+     */
+    public function isDevEnvironment(): bool {
+        return $this->environment == self::ENVIRONMENT_DEV;
+    }
+
+    /**
+     * Returns Layout based on provided layout name
+     * @var string $layout
+     * @return Layout
+     */
+    public function setLayoutByName(string $layout): Layout {
+        // Legacy layout support
+        $system = $this;
+
+        switch($layout) {
+            case 'cextralite':
+                $this->layout = require "layout/cextralite.php";
+                break;
+            case 'classic_blue':
+                $this->layout = require  "layout/classic_blue.php";
+                break;
+            case 'shadow_ribbon':
+                $this->layout = require  "layout/shadow_ribbon.php";
+                break;
+            case 'geisha':
+                $this->layout = require  "layout/geisha.php";
+                break;
+            case 'blue_scroll':
+                $this->layout = require  "layout/blue_scroll.php";
+                break;
+            case 'rainbow_road':
+                $this->layout = require "layout/rainbow_road.php";
+                break;
+            case 'new_geisha':
+                require_once __DIR__ . "/../layout/new_geisha.php";
+                // This needs to be first so the function can read it
+                $this->enable_mobile_layout = true;
+                $this->layout = getNewGeishaLayout($this, $this->enable_mobile_layout);
+                break;
+            case 'sumu':
+                require_once __DIR__ . "/../layout/sumu.php";
+                // This needs to be first so the function can read it
+                $this->enable_mobile_layout = true;
+                $this->layout = getSumuLayout($this, $this->enable_mobile_layout);
+                break;
+            default:
+                $this->layout = require "layout/" . self::DEFAULT_LAYOUT . ".php";
+                break;
+        }
+
+        return $this->layout;
+    }
+
+    /**
+     * Checks for active events and sets weekend & holiday boosts
+     * By default, this will use Server time
+     * @return void
+     */
+    public function checkForActiveEvent(): void {
+        $SERVER_TIME_ZONE = new DateTimeZone(self::SERVER_TIME_ZONE);
+        // Set events on development environments
+        if($this->isDevEnvironment()) {
+            $july_2023_lantern_event_start_time = new DateTimeImmutable('2023-07-15', $SERVER_TIME_ZONE);
+            $double_exp_start_time = new DateTimeImmutable('2023-12-25', $SERVER_TIME_ZONE);
+            $double_reputation_start_time = new DateTimeImmutable('2024-01-19', $SERVER_TIME_ZONE);
+        }
+        // Set events on production environments
+        else {
+            $july_2023_lantern_event_start_time = new DateTimeImmutable('2023-07-15', $SERVER_TIME_ZONE);
+            $double_exp_start_time = new DateTimeImmutable('2023-12-25', $SERVER_TIME_ZONE);
+            $double_reputation_start_time = new DateTimeImmutable('2023-10-16', $SERVER_TIME_ZONE);
+        }
+
+        /********* CORE EVENTS *********/
+        // Double exp gains
+        $double_exp_end_time = new DateTimeImmutable('2023-12-27');
+        if($this->SERVER_TIME > $double_exp_start_time && $this->SERVER_TIME < $double_exp_end_time) {
+            $this->event = new DoubleExpEvent($double_exp_end_time);
+        }
+        // Double reputation gains
+        $double_reputation_end_time = new DateTimeImmutable('2024-01-2');
+        if($this->SERVER_TIME > $double_reputation_start_time && $this->SERVER_TIME < $double_reputation_end_time) {
+            $this->event = new DoubleReputationEvent($double_reputation_end_time);
+        }
+
+        /***** LIMITED TIME EVENTS *****/
+        $july_2023_lantern_event_end_time = new DateTimeImmutable('2023-07-16', $SERVER_TIME_ZONE);
+        if($this->SERVER_TIME > $july_2023_lantern_event_start_time && $this->SERVER_TIME < $july_2023_lantern_event_end_time) {
+            $this->event = new LanternEvent($july_2023_lantern_event_end_time);
+        }
+
+        /******* HOLIDAY BOOSTS ********/
+        if (!isset($this->event) && in_array($this->SERVER_TIME->format('M j'), array_keys(self::HOLIDAYS))) {
+            $endTime = new DateTimeImmutable("tomorrow", $SERVER_TIME_ZONE);
+            $this->event = new HolidayBonusEvent($endTime, self::HOLIDAYS[$this->SERVER_TIME->format('M j')]);
+        }
+
+        /******* WEEKEND BOOSTS ********/
+        if (!isset($this->event) && (System::currentDayOfWeek() == 0 || System::currentDayOfWeek() == 6)) {
+            $endTime = new DateTimeImmutable('next Monday', $SERVER_TIME_ZONE);
+            $this->event = new BonusExpWeekend($endTime);
+        }
+
+        // This needs to be last - test notification on DEV servers
+        if($this->testNotifications['event'] && is_null($this->event) && $this->isDevEnvironment()) {
+            $this->event = new DoubleExpEvent($this->SERVER_TIME->modify("+2 weeks"));
+        }
+    }
+
+    /**
+     * Closes and opens server in-real time, without needing manual code edits
+     * Sets maintenance time-frame and displays a countdown on non-legacy layouts
+     * @return void
+     */
+    public function checkForMaintenance(): void {
+        // Check for db-based maintenance
+        $db_maint = $this->db->query("SELECT `maintenance_begin_time`, `maintenance_end_time` FROM `system_storage`");
+        $db_maint = $this->db->fetch($db_maint);
+
+        // Hard closure
+        if($db_maint['maintenance_end_time'] == -1) {
+            $this->SC_OPEN = false;
+            $this->UPDATE_MAINTENANCE = null;
+        }
+
+        if($db_maint['maintenance_begin_time'] > $this->SERVER_TIME->getTimestamp() || $db_maint['maintenance_end_time'] > $this->SERVER_TIME->getTimestamp()) {
+            $maintenanceBegin = $this->SERVER_TIME->setTimestamp($db_maint['maintenance_begin_time']);
+            $maintenanceEnd = $this->SERVER_TIME->setTimestamp($db_maint['maintenance_end_time']);
+
+            // Display timer for maintenance window
+            if($this->SERVER_TIME->getTimestamp() < $maintenanceBegin->getTimestamp()) {
+                $this->UPDATE_MAINTENANCE = $maintenanceBegin;
+            }
+            // Close SC for maintenance window - NOTE: This can be overridden in vars.php if window can't be easily determined
+            if($this->SERVER_TIME->getTimestamp() > $maintenanceBegin->getTimestamp() && $this->SERVER_TIME->getTimestamp() < $maintenanceEnd->getTimestamp()) {
+                $this->SC_OPEN = false;
+                $this->UPDATE_MAINTENANCE = $maintenanceEnd;
+            }
+        }
+    }
+
+    /**
+     * Returns estimated time server will come back online, rounded up to nearst 5 minutes
+     * E.g. 7 minutes => 10 minutes // 1 minute => 5 minutes
+     * If system is closed, returns generic closed message
+     * @return string
+     */
+    public function getMaintenenceEndTime(): string {
+        if(!$this->SC_OPEN && $this->UPDATE_MAINTENANCE) {
+            $mins = ceil(($this->UPDATE_MAINTENANCE->getTimestamp() - time()) / 60); // Round up to nearest minute
+            $mins += 5 - ($mins % 5); // Add remainder to bring to nearst 5 minutes
+            return "$mins minutes";
+        }
+
+        return "1 hour";
+    }
+
+    /**
+     * Returns entity based on type & ID
+     * @param string $entity_id
+     * @return EntityId
+     * @throws RuntimeException
+     */
+    public static function parseEntityId(string $entity_id): EntityId {
+        $arr = explode(':', $entity_id);
+        if(count($arr) != 2) {
+            throw new RuntimeException("Invalid entity id {$entity_id}!");
+        }
+
+        return new EntityId($arr[0], (int)$arr[1]);
+    }
+
+    /**
+     * Returns build file for react components
+     * @var string $component_name
+     * @return string
+     */
     public function getReactFile(string $component_name): string {
         $filename = "ui_components/build/{$component_name}.js";
         return $this->router->base_url . $filename . "?v=" .  filemtime($filename);
     }
 
+    /**
+     * Returns src string of css file
+     * @var string $file_name
+     * @return string
+     */
     public function getCssFileLink(string $file_name): string {
         return $this->router->base_url . $file_name . "?v=" .  filemtime($file_name);
     }
 
-    public static function clampNumber($number, $min, $max)
-    {
+    /**
+     * Helper function
+     * @return int
+     */
+    public static function clampNumber($number, $min, $max): int {
         return max(min($number, $max), $min);
     }
 
@@ -1039,5 +1125,88 @@ class SystemOld {
         parse_str(implode('&', array_slice($_SERVER['argv'], 1)), $result);
 
         return $result;
+    }
+
+    /**
+     * Initializes and returns system
+     * @param bool $api_request
+     * @return System
+     * @throws Exception
+     */
+    public static function initialize(bool $api_request = false): System {
+        /**
+         * This must be called here to properly pull variables into initializer
+         * @var $host
+         * @var $username
+         * @var $password
+         * @var $database
+         *
+         * @var string $ENVIRONMENT
+         * @var string $WEB_URL
+         * @var bool $SC_OPEN
+         * @var bool $ENABLE_DEV_ONLY_FEATURES
+         * @var bool $LOCAL_HOST_CONNECTION
+         * @var bool $REGISTER_OPEN
+         * @var bool $USE_NEW_BATTLES
+         * @var bool $WAR_ENABLED
+         * @var bool $REQUIRE_USER_VERIFICATION
+         */
+        require_once __DIR__ . '/../secure/vars.php';
+
+        $system = new System(
+            db: new Database($host, $username, $password, $database),
+            router: new Router($WEB_URL),
+            SC_OPEN: $SC_OPEN,
+            USE_NEW_BATTLES: $USE_NEW_BATTLES,
+            war_enabled: $WAR_ENABLED,
+            REQUIRE_USER_VERIFICATION: $REQUIRE_USER_VERIFICATION,
+            SERVER_TIME: new DateTimeImmutable("now", new DateTimeZone(self::SERVER_TIME_ZONE)),
+            environment: $ENVIRONMENT,
+            enable_dev_only_features: $ENABLE_DEV_ONLY_FEATURES,
+            local_host: $LOCAL_HOST_CONNECTION,
+            register_open: $REGISTER_OPEN
+        );
+
+        // Load reputation layout, reputation reset and check for server maintenance
+        $system->loadRepReset();
+        $system->checkForActiveEvent();
+        $system->checkForMaintenance();
+
+        // Set home page variables
+        $system->homeVars = [
+            'view' => 'none',
+            // TODO: This can be removed, header API has the two links used available
+            'links' => [
+                'news_api' => $system->router->api_links['news'],
+                'logout' => $system->router->base_url . '?logout=1',
+                'profile' => $system->router->getUrl('profile'),
+                'github' => $system->router->links['github'],
+                'discord' => $system->router->links['discord'],
+                'support' => $system->router->base_url . 'support.php',
+                'login_url' => $system->router->base_url,
+                'register_url' => $system->router->base_url,
+            ],
+            'errors' => [
+                'login' => '',
+                'register' => '',
+                'reset' => '',
+            ],
+            'messages' => [
+                'login' => '',
+                'register' => '',
+                'reset' => '',
+            ],
+            'register_prefill' => [],
+        ];
+
+        // Load layout if this is not an api request
+        if(!$api_request) {
+            $system->setLayoutByName(self::DEFAULT_LAYOUT);
+        }
+
+        // Legacy layout support
+        $system->timezoneOffset = date(format: 'Z');
+
+        return $system;
     }
 }
