+++ cron_jobs/caravan_cron.php
@@ -16,7 +16,7 @@ session_start();
  *      Creates resource logs
  */
 
-require_once __DIR__ . '/../classes/SystemV2.php';
+require_once __DIR__ . '/../classes/System.php';
 require_once __DIR__ . '/../classes/Village.php';
 require_once __DIR__ . '/../classes/User.php';
 require_once __DIR__ . '/../classes/war/WarManager.php';
