+++ cron_jobs/monthly_reset_cron.php
@@ -1,7 +1,7 @@
 <?php
 session_start();
 
-require_once __DIR__ . '/../classes/SystemV2.php';
+require_once __DIR__ . '/../classes/System.php';
 require_once __DIR__ . '/../classes/User.php';
 
 /**
