+++ cron_jobs/region_cron.php
@@ -14,7 +14,7 @@ session_start();
  *      Creates resource logs
  */
 
-require_once __DIR__ . '/../classes/SystemV2.php';
+require_once __DIR__ . '/../classes/System.php';
 require_once __DIR__ . '/../classes/Village.php';
 require_once __DIR__ . '/../classes/User.php';
 require_once __DIR__ . '/../classes/war/WarManager.php';
