+++ cron_jobs/reputation_cron.php
@@ -11,7 +11,7 @@ session_start();
  *    will also remove players with insufficient reputation from office.    *
  ****************************************************************************/
 
-require_once __DIR__ . '/../classes/SystemV2.php';
+require_once __DIR__ . '/../classes/System.php';
 require_once __DIR__ . '/../classes/Village.php';
 require_once __DIR__ . '/../classes/User.php';
 require_once __DIR__ . '/../classes/UserReputation.php';
